+++ firrtl/src/main/scala/firrtl/analyses/InstanceKeyGraph.scala
@@ -130,11 +130,8 @@ object InstanceKeyGraph {
     case ir.Module(_, _, _, body) => {
       val instances = mutable.ArrayBuffer[InstanceKey]()
       def onStmt(s: ir.Statement): Unit = s match {
-        case firrtl.WDefInstance(_, name, module, _) => instances += InstanceKey(name, module)
-        case ir.DefInstance(_, name, module, _)      => instances += InstanceKey(name, module)
-        case _: firrtl.WDefInstanceConnector =>
-          firrtl.Utils.throwInternalError("Expecting WDefInstance, found a WDefInstanceConnector!")
-        case other => other.foreachStmt(onStmt)
+        case ir.DefInstance(_, name, module, _) => instances += InstanceKey(name, module)
+        case other                              => other.foreachStmt(onStmt)
       }
       onStmt(body)
       instances.toSeq
